+++ 99reddits/iPad/AlbumView/AlbumViewLayoutPad.m
@@ -45,13 +45,13 @@
 - (UICollectionViewLayoutAttributes *)layoutAttributesForItemAtIndexPath:(NSIndexPath *)indexPath {
     UICollectionViewLayoutAttributes *attributes = [UICollectionViewLayoutAttributes layoutAttributesForCellWithIndexPath:indexPath];
 	if (UIInterfaceOrientationIsLandscape([[UIApplication sharedApplication] statusBarOrientation])) {
-		int row = indexPath.item / 7;
-		int col = indexPath.item % 7;
+		NSInteger row = indexPath.item / 7;
+		NSInteger col = indexPath.item % 7;
 		attributes.frame = CGRectMake(23 + col * 143, 15 + row * 150, 120, 120);
 	}
 	else {
-		int row = indexPath.item / 5;
-		int col = indexPath.item % 5;
+		NSInteger row = indexPath.item / 5;
+		NSInteger col = indexPath.item % 5;
 		attributes.frame = CGRectMake(28 + col * 148, 15 + row * 150, 120, 120);
 	}
     return attributes;
@@ -92,7 +92,7 @@
 	if (attributes == nil)
 		attributes = [self layoutAttributesForItemAtIndexPath:itemIndexPath];
 
-	int count = 0;
+	NSInteger count = 0;
 	for (NSIndexPath *indexPath in insertIndexPaths) {
 		if (indexPath.row < itemIndexPath.row)
 			count --;
@@ -104,16 +104,16 @@
 
 	if ([insertIndexPaths containsObject:itemIndexPath]) {
 		if (UIInterfaceOrientationIsLandscape([[UIApplication sharedApplication] statusBarOrientation])) {
-			int row = itemIndexPath.item / 7;
-			int col = itemIndexPath.item % 7;
+			NSInteger row = itemIndexPath.item / 7;
+			NSInteger col = itemIndexPath.item % 7;
 			if (isIOS7Below)
 				attributes.frame = CGRectMake(23 + col * 143, 15 + row * 150, 120, 120);
 			else
 				attributes.frame = CGRectMake(23 + col * 143, 79 + row * 150, 120, 120);
 		}
 		else {
-			int row = itemIndexPath.item / 5;
-			int col = itemIndexPath.item % 5;
+			NSInteger row = itemIndexPath.item / 5;
+			NSInteger col = itemIndexPath.item % 5;
 			if (isIOS7Below)
 				attributes.frame = CGRectMake(28 + col * 148, 15 + row * 150, 120, 120);
 			else
@@ -125,16 +125,16 @@
 	}
 	else {
 		if (UIInterfaceOrientationIsLandscape([[UIApplication sharedApplication] statusBarOrientation])) {
-			int row = (itemIndexPath.item + count) / 7;
-			int col = (itemIndexPath.item + count) % 7;
+			NSInteger row = (itemIndexPath.item + count) / 7;
+			NSInteger col = (itemIndexPath.item + count) % 7;
 			if (isIOS7Below)
 				attributes.frame = CGRectMake(23 + col * 143, 15 + row * 150, 120, 120);
 			else
 				attributes.frame = CGRectMake(23 + col * 143, 79 + row * 150, 120, 120);
 		}
 		else {
-			int row = (itemIndexPath.item + count) / 5;
-			int col = (itemIndexPath.item + count) % 5;
+			NSInteger row = (itemIndexPath.item + count) / 5;
+			NSInteger col = (itemIndexPath.item + count) % 5;
 			if (isIOS7Below)
 				attributes.frame = CGRectMake(28 + col * 148, 15 + row * 150, 120, 120);
 			else
@@ -152,16 +152,16 @@
 		attributes = [self layoutAttributesForItemAtIndexPath:itemIndexPath];
 
 	if (UIInterfaceOrientationIsLandscape([[UIApplication sharedApplication] statusBarOrientation])) {
-		int row = itemIndexPath.item / 7;
-		int col = itemIndexPath.item % 7;
+		NSInteger row = itemIndexPath.item / 7;
+		NSInteger col = itemIndexPath.item % 7;
 		if (isIOS7Below)
 			attributes.frame = CGRectMake(23 + col * 143, 15 + row * 150, 120, 120);
 		else
 			attributes.frame = CGRectMake(23 + col * 143, 79 + row * 150, 120, 120);
 	}
 	else {
-		int row = itemIndexPath.item / 5;
-		int col = itemIndexPath.item % 5;
+		NSInteger row = itemIndexPath.item / 5;
+		NSInteger col = itemIndexPath.item % 5;
 		if (isIOS7Below)
 			attributes.frame = CGRectMake(28 + col * 148, 15 + row * 150, 120, 120);
 		else
