+++ 99reddits/iPad/MainView/MainViewCellPad.h
@@ -27,8 +27,8 @@
 	UILabel *nameLabel;
 	UIImageView *animateImageView;
 
-	int unshowedCount;
-	int totalCount;
+	NSInteger unshowedCount;
+	NSInteger totalCount;
 	BOOL loading;
 	
 	BOOL bFavorites;
@@ -41,8 +41,8 @@
 @property (nonatomic, weak) SubRedditItem *subReddit;
 @property (nonatomic, readonly) UILabel *nameLabel;
 
-- (void)setUnshowedCount:(int)_unshowedCount totalCount:(int)_totalCount loading:(BOOL)_loading;
-- (void)setTotalCount:(int)_totalCount;
+- (void)setUnshowedCount:(NSInteger)_unshowedCount totalCount:(NSInteger)_totalCount loading:(BOOL)_loading;
+- (void)setTotalCount:(NSInteger)_totalCount;
 - (void)setThumbImage:(UIImage *)thumbImage animated:(BOOL)animated;
 - (void)setEditing:(BOOL)_editing;
