+++ 99reddits/iPhone/MainView/MainViewController.m
@@ -375,7 +375,7 @@
 	
 	refreshCount = 0;
 	
-	for (int i = 0; i < subRedditsArray.count; i ++) {
+	for (NSInteger i = 0; i < subRedditsArray.count; i ++) {
 		SubRedditItem *subReddit = [subRedditsArray objectAtIndex:i];
 		subReddit.loading = YES;
 		
@@ -397,8 +397,8 @@
 	NSString *urlString = [[request originalURL] absoluteString];
 
 	SubRedditItem *subReddit = nil;
-	int index = 0;
-	for (int i = 0; i < subRedditsArray.count; i ++) {
+	NSInteger index = 0;
+	for (NSInteger i = 0; i < subRedditsArray.count; i ++) {
 		SubRedditItem *tempSubReddit = [subRedditsArray objectAtIndex:i];
 		if ([tempSubReddit.urlString isEqualToString:urlString]) {
 			subReddit = tempSubReddit;
@@ -451,8 +451,8 @@
 	NSString *urlString = [[request originalURL] absoluteString];
 	
 	SubRedditItem *subReddit = nil;
-	int index = 0;
-	for (int i = 0; i < subRedditsArray.count; i ++) {
+	NSInteger index = 0;
+	for (NSInteger i = 0; i < subRedditsArray.count; i ++) {
 		SubRedditItem *tempSubReddit = [subRedditsArray objectAtIndex:i];
 		if ([tempSubReddit.urlString isEqualToString:urlString]) {
 			subReddit = tempSubReddit;
@@ -624,12 +624,12 @@
 	[readOp setCompletionBlock:^{
 		UIImage *image = [UIImage imageWithData:[readOp responseData]];
 		
-		int index = -2;
+		NSInteger index = -2;
 		if (photoIndex == -1) {
 			index = -1;
 		}
 		else {
-			for (int i = 0; i < subRedditsArray.count; i ++) {
+			for (NSInteger i = 0; i < subRedditsArray.count; i ++) {
 				NSString *keyString = [self cacheKeyForPhotoIndex:i];
 				if ([keyString isEqualToString:photoIndexKey]) {
 					index = i;
@@ -640,7 +640,7 @@
 		
 		if (index != -2) {
 			if (image && (subRedditsArray.count + 1 > photoIndex || photoIndex == -1)) {
-				int x, y, w, h;
+				NSInteger x, y, w, h;
 				if (image.size.width > THUMB_WIDTH * 2 && image.size.height > THUMB_HEIGHT * 2) {
 					float imgRatio = image.size.width / image.size.height;
 					if (imgRatio < 1) {
